Make access to CPU state explicit
Having explicit accesses to the CPU state is currently mandatory for
usage as LLVM helpers, or an abort will be emitted. This is done to
ensure that we know precisely which parts of the CPU state a set of
helpers access.

With a more sophisticated analysis it's possible to avoid having to do
this ugly hack, which make merging with upstream harder.