Ensure cc_op is always serialized
As with x86, QEMU uses a set of CPU state fields (`cc_*`) to avoid
computing condition flags each time an operation that affects them is
performed.

This can lead to some issues in rev.ng due to the fact that we split a
basic block that has been optimized by QEMU making some assumptions. The
lower part of the block is (correctly re-created) but upper part is left
as is.

In certain situations, due to this fact, we end up missing an
instruction updating `cc_op`, the field containing the identifier of the
last operation that affected the condition flags.

This commit ensures `cc_op` is always updated correctly.