Fix generation of exit_tb in gen_tb_end()
This commit fixes the generation of the final instructions of a
`TranslationBlock` when `CONFIG_LIBTINYCODE` is defined.

Before this commit the `gen_tb_end()` function did nothing when
`CONFIG_LIBTINYCODE` was defined. For this reason some blocks generated
with libtinycode were not properly terminated by a `exit_tb` tinycode
instruction. This in turn caused problems in libtinycode's users,
because they expected to use `exit_tb` as markers for the end of a
`TranslationBlock`.

This commit fixes this situation, emitting an `exit_tb` tinycode
instruction at the end of each `TranslationBlock` if the block is not
already properly terminated.