Fix wrong `#ifdef` in MIPS delay slots handling
In MIPS, we disabled emission of the `debug_insn` pseudo instruction in
case an instruction is in a delay slot. This was required to consider
the delay slot as "part of" the branch instruction.

However, we were doing this only if `LLVM_HELPERS` was defined, while
this was required by libtinycode. As a result, the change had no effect.

This commit fixes the `#ifdef`.