llvm-helpers: disable certain features
Ideally, this commit should not exist.
We should move this logic to use attributes that rev.ng can later
post-process.

Some of these macros are in "common" code. As long as we have this, we
cannot have a single build (as in invocation of `./configure`) for all
the targets needed by rev.ng.